Upload the finished project
I had Miniconda in my folder I don't think it is doing anything, but it is not in my GitHub because it is too large. I figure it wasn't crucial, but just in case it was making my environment work, I thought I would make a note of it. I think it was one of those Copilot fixes that didn't do anything. Haha.